+++ libs/web/src/grids/todo-statuses-grid/todo-statuses-grid.component.ts
@@ -14,6 +14,7 @@ import { TodoStatus } from '@rucken/todo-core';
 import { ShortTodoProject } from '@rucken/todo-core';
 import { BaseResourcesGridComponent } from '@rucken/web';
 import { ConfirmModalComponent } from '@rucken/web';
+import { takeUntil } from 'rxjs/operators';
 
 import { TodoStatusModalComponent } from './todo-status-modal/todo-status-modal.component';
 
@@ -69,7 +70,7 @@ export class TodoStatusesGridComponent extends BaseResourcesGridComponent {
     itemModal.item.project = this.project;
     itemModal.modal.show();
     this.selectedItems = [itemModal.item];
-    this.cachedResourcesService.changeStatusItem$.takeUntil(this.destroyed$).subscribe(status =>
+    this.cachedResourcesService.changeStatusItem$.pipe(takeUntil(this.destroyed$)).subscribe((status: any) =>
       itemModal.okInProcessFromStatus(status)
     );
   }
@@ -90,7 +91,7 @@ export class TodoStatusesGridComponent extends BaseResourcesGridComponent {
     itemModal.item = new TodoStatus(item);
     itemModal.modal.show();
     this.selectedItems = [itemModal.item];
-    this.cachedResourcesService.changeStatusItem$.takeUntil(this.destroyed$).subscribe(status =>
+    this.cachedResourcesService.changeStatusItem$.pipe(takeUntil(this.destroyed$)).subscribe((status: any) =>
       itemModal.okInProcessFromStatus(status)
     );
   }
@@ -107,7 +108,7 @@ export class TodoStatusesGridComponent extends BaseResourcesGridComponent {
     confirm.onClose.subscribe(() => this.focus());
     this.selectedItems = [item];
     confirm.modal.show();
-    this.cachedResourcesService.changeStatusItem$.takeUntil(this.destroyed$).subscribe(status =>
+    this.cachedResourcesService.changeStatusItem$.pipe(takeUntil(this.destroyed$)).subscribe((status: any) =>
       confirm.okInProcessFromStatus(status)
     );
   }