+++ srcgen/grid.select.input.modal/libs/web/src/grids/%grid.list.name.kebab%-grid/%grid.list.name.kebab%-list-modal/%grid.list.name.kebab%-list-modal.module.ts
@@ -1,7 +1,7 @@
 import { <%=grid.list.name.camel%>GridModule } from '../<%=grid.list.name.kebab%>-grid.module';
 import { NgModule, ModuleWithProviders } from '@angular/core';
 import { <%=grid.list.name.camel%>ListModalComponent } from './<%=grid.list.name.kebab%>-list-modal.component';
-import { ModalModule } from 'ngx-bootstrap';
+import { ModalModule } from 'ngx-bootstrap/modal';
 import { SharedModule } from '@rucken/web';
 import { FooterButtonsModule } from '@rucken/web';